+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.2 March 15, 2024
 
 - Bug #9: Fix broken links in `Div`, `Li`, `Ol`, `P`, and `Ul` documentation (@terabytesoftw)
+- Enh #10: Add `Hr` class widget (@terabytesoftw)
 
 ## 0.1.1 March 8, 2024